front/requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepmerge==2.0
Deprecated==1.2.15
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.6.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.16.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pip==22.3
pip-requirements-parser==32.0.1
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
schema==0.7.7
scipy==1.14.1
seaborn==0.13.2
setuptools==65.5.0
simple-di==0.1.5
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
tomli_w==1.1.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.7
uvicorn==0.32.1
watchdog==6.0.0
watchfiles==1.0.0
wcwidth==0.2.13
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0